Bumping versions of dependencies
Test vectors of encryption methods are updated. StdRng was changed from ChaCha20 to ChaCha12 in the new version of `rand`, which leads to a change in seed generation. It's okay since decryption works in a backward-compatible manner as expected. Also, it is actually expected to have a random seed in a wild, so tests just unveiled the issue which is good.